# src/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "descriptuon": "sample server",
        "version": "1.0.0",
        "title": "Assignment Test",
        "contact": {
            "email": "arun.ang.js@gmail.com"
        }
    },
    "host": "localhost:8000",
    "basePath": "/api",
...
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/admin": {
            "post": {
                "tags": ["Admin"],
                "Summary": "Create a Admin data",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "The authorization token",
                        "required": true,
                        "schema": {
                          "type": "string"
                        }
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/createAdmin"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "integer",
                                    "default": "0"
                                },
                                "message": {
                                    "type": "string",
                                    "default": "The password must contain a lowercase letter, an uppercase letter, a number, a special character, minimum 8 characters and maximum 20 character !! \"or\" This Email already exists. Try giving another email !! \"or\" Unable to create the Admin !!"
                                }
                            }
                        }
                    },
                    "200": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "integer",
                                    "default": "1"
                                },
                                "message": {
                                    "type": "string",
                                    "default": "Successfully created the Admin !!"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": ["Admin"],
                "Summary": "Create a Admin data",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "The authorization token",
                        "required": true,
                        "schema": {
                          "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "integer",
                                    "default": "1"
                                },
                                "message": {
                                    "type": "string",
                                    "default": "Successfully got the admin list !!"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/admin/super-login": {
            "post": {
                "tags": ["Login"],
                "Summary": "Create a Feed data",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/login"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Error Message",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "integer",
                                    "default": "0"
                                },
                                "message": {
                                    "type": "string",
                                    "default": "Unable to create the Feed !!"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success Message",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "integer",
                                    "default": "1"
                                },
                                "message": {
                                    "type": "string",
                                    "default": "Successfully created the Feed !!"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/admin/{id}": {
            "put": {
                "tags": ["Admin"],
                "Summary": "Update a admin data",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "The authorization token",
                        "required": true,
                        "schema": {
                          "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/update"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Error Message",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "integer",
                                    "default": "0"
                                },
                                "message": {
                                    "type": "string",
                                    "default": "Invalid admin id !! \" or \" Unable to update the admin data !!"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success Message",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "integer",
                                    "default": "1"
                                },
                                "message": {
                                    "type": "string",
                                    "default": "Successfully update the admin data !!"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": ["Admin"],
                "Summary": "Get a detail of admin data",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "The authorization token",
                        "required": true,
                        "schema": {
                          "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Error Message",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "integer",
                                    "default": "0"
                                },
                                "message": {
                                    "type": "string",
                                    "default": "Invalid admin id !!"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success Message",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "integer",
                                    "default": "1"
                                },
                                "message": {
                                    "type": "string",
                                    "default": "Successfully got the detail of a admin !!"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": ["Admin"],
                "Summary": "Delete an admin data",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "The authorization token",
                        "required": true,
                        "schema": {
                          "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Error Message",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "integer",
                                    "default": "0"
                                },
                                "message": {
                                    "type": "string",
                                    "default": "Invalid admin id !! \" or \" Unable to delet an admin !!"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success Message",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "integer",
                                    "default": "1"
                                },
                                "message": {
                                    "type": "string",
                                    "default": "Successfully delete an admin !!"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/admin/log/file-list": {
            "get": {
                "tags": ["Log-List"],
                "Summary": "Create a user data",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "The authorization token",
                        "required": true,
                        "schema": {
                          "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "integer",
                                    "default": "1"
                                },
                                "message": {
                                    "type": "string",
                                    "default": "Successfully got the user list !!"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "createAdmin": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "number",
                    "format": "int-32"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "login": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "update": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "role": {
                    "type": "integer",
                    "format": "int-32"
                }
            }
        },
...
    }
}